Moves render_target protocol to protocols package
