+++ src/pyrite/_types/protocols.py
@@ -4,7 +4,7 @@ from typing import Protocol, TYPE_CHECKING
 
 
 if TYPE_CHECKING:
-    from pygame import Surface, Vector2
+    from pygame import Rect, Surface, Vector2
     from pyrite._transform.transform_component import TransformComponent
 
 
@@ -62,3 +62,26 @@ class CanPostUpdate(Protocol):
 class CanConstUpdate(Protocol):
 
     def const_update(self, timestep: float) -> None: ...
+
+
+class RenderTarget(Protocol):
+
+    def get_target_surface(self) -> Surface:
+        """
+        Gets the surface to be drawn to.
+        """
+        ...
+
+    def get_target_rect(self) -> Rect:
+        """
+        Gets the Rect that represents the surface space of the RenderTarget.
+        """
+        ...
+
+    @property
+    def crop(self) -> bool:
+        """
+        Determines if the rendering should be cropped or not.
+        If False, the rendering will be scaled to fit the target.
+        """
+        ...
